% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{c_mtp}
\alias{c_mtp}
\title{Multiple testing following the graph}
\usage{
c_mtp(p_values, alphas, mat_g)
}
\arguments{
\item{p_values}{matrix of p-values for the elementary hypothesis}

\item{alphas}{vector of the original alphas}

\item{mat_g}{transition matrix G}
}
\value{
Hypothesis rejection status indicator vector
}
\description{
Multiple testing following the graph
}
