% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{c_mtp_single}
\alias{c_mtp_single}
\title{Multiple testing following the graph}
\usage{
c_mtp_single(p_values, alphas, mat_g, log = FALSE)
}
\arguments{
\item{p_values}{vector of p-values for the elementary hypothesis}

\item{alphas}{vector of the original alphas}

\item{mat_g}{transition matrix G}

\item{log}{TRUE: print log at each step; FALSE: silent}
}
\value{
Hypothesis rejection status indicator vector
}
\description{
Multiple testing following the graph
}
