% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/om_present.R
\name{om_print_grp}
\alias{om_print_grp}
\title{Print latex graph}
\usage{
om_print_grp(
  mat_g,
  weights,
  fname = "tmp.org",
  pre_org = "#+LATEX_HEADER: \\\\usepackage{tikz} \\n\\n",
  pre = "\\\\begin{figure}\\\\resizebox{0.8\\\\textwidth}{!}{\%\\n",
  post = "\\n}\\n\\\\end{figure} \\n"
)
}
\description{
Print latex graph
}
