% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/om_optim.R
\name{om_rejection}
\alias{om_rejection}
\title{Compute rejection target function}
\usage{
om_rejection(
  weights,
  mat_g,
  p_values,
  alpha = 0.05,
  utility = 1,
  null_h = NULL
)
}
\arguments{
\item{weights}{vector of the original weights}

\item{mat_g}{transition matrix G}

\item{p_values}{vector of p-values for the elementary hypotheses}

\item{alpha}{overall alpha level}

\item{utility}{rejection utility of each elementary hypothesis}

\item{null_h}{null hypothesis index}

\item{theta}{parameters in weights and G}
}
\value{
vector of the following values: 1. rejection utility, by default,
    this is the average number of rejected hypothesis 2. reject any null
    hypothesis rate, i.e., familywise false discovery rate 3. average
    rejection rate for each hypothesis
}
\description{
Compute rejection target function
}
