% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/om_optim.R
\name{om_rejection_optim}
\alias{om_rejection_optim}
\title{Optimize weights and G}
\usage{
om_rejection_optim(
  weights,
  mat_g,
  ...,
  tot = 1,
  init_theta = NULL,
  uncons_theta = TRUE,
  par_optim = list()
)
}
\arguments{
\item{weights}{vector of the original weights}

\item{mat_g}{transition matrix G}

\item{tot}{restrict on the sum of parameters}

\item{init_theta}{initial values for theta. Default NULL, which will use
00000.1 fro all theta as initial values.}

\item{uncons_theta}{whether theta are unconstrained. If TRUE, convert using
exponential function. Example: (row tot - row sum fixed values)
[e^a/(1+e^a+e^b) e^b/(1+e^a+e^b)]}

\item{par_optim}{options for constrOptim function}
}
\description{
Optimize weights and G
}
\details{
Parameters in weights and mat_g should be entered as NA. The
    constraints require all the parameter to be non-negative and each row in
    mat_g and weights have total 1.

    Note that the first NA in each row of mat_g or weights are fixed values
    given the other elements in the row in order to satisfy the constraint
    that the sum is 1.
}
