% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/om_simu.R
\name{om_simu}
\alias{om_simu}
\title{Simulate endpoint}
\usage{
om_simu(rep, s_mu, s_sd, s_cor, s_size = 1, seed = NULL)
}
\arguments{
\item{rep}{number of replications}

\item{s_mu}{simulation mean}

\item{s_sd}{simulation SD (i.e. individual subject)}

\item{s_cor}{correlation (i.e. individual subject)}

\item{s_size}{sample size}
}
\value{
Matrix of simulated test statistics
}
\description{
Simulate endpoint
}
