% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/om_optim.R
\name{om_theta_fill}
\alias{om_theta_fill}
\title{Fill parameters to the G and weights}
\usage{
om_theta_fill(theta = NULL, weights, mat_g, uncons_theta = TRUE, tot = 1)
}
\arguments{
\item{theta}{parameters in weights and G}

\item{weights}{vector of the original weights}

\item{mat_g}{transition matrix G}

\item{uncons_theta}{whether theta are unconstrained. If TRUE, convert using
exponential function. Example: (row tot - row sum fixed values)
[e^a/(1+e^a+e^b) e^b/(1+e^a+e^b)]}
}
\description{
Fill parameters to the G and weights
}
