% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/om_simu.R
\name{om_z_pval}
\alias{om_z_pval}
\title{Get P-values}
\usage{
om_z_pval(
  stats,
  null_mu = 0,
  null_sd = 1,
  type = c("two-sided", "left", "right")
)
}
\arguments{
\item{stats}{simulated test statistics}

\item{null_mu}{mu under NULL}

\item{s_sd}{simulation SD}

\item{s_cor}{correlation}
}
\value{
p-values
}
\description{
Get P-values
}
