% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/si_bayesian.R
\name{si_bayes_beta_post}
\alias{si_bayes_beta_post}
\title{Get posterior distribution}
\usage{
si_bayes_beta_post(
  x = seq(0, 1, by = 0.01),
  obs_y = 1,
  obs_n = 10,
  pri_ab = c(1, 1)
)
}
\description{
Get posterior distribution
}
