% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/si_bayesian.R
\name{si_bayes_hierarchy}
\alias{si_bayes_hierarchy}
\title{Bayesian hierarchical modeling}
\usage{
si_bayes_hierarchy(
  vec_y,
  vec_n,
  ...,
  pri_sig = 2.5,
  adjust = 1.5,
  by_study = FALSE,
  pri_ab = c(0.5, 0, 5),
  label = "Subgroup"
)
}
\description{
Bayesian hierarchical modeling
}
