% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/si_bayesian.R
\name{si_bayes_plot_tipping}
\alias{si_bayes_plot_tipping}
\title{Plot tippling point analysis result}
\usage{
si_bayes_plot_tipping(mat_tipping, f_col = NULL, add_pval = TRUE)
}
\description{
Plot tippling point analysis result
}
