% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/si_boundary.R
\name{si_bd_get_bd}
\alias{si_bd_get_bd}
\title{Get Boundary}
\usage{
si_bd_get_bd(
  type = c("NO", "P", "OF", "HP", "asP", "asOF"),
  alpha = 0.025,
  n_interim = 1,
  info_frac = NULL
)
}
\description{
Get Boundary
}
