% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/si_boundary.R
\name{si_bd_plt_rep}
\alias{si_bd_plt_rep}
\title{Plot data for all}
\usage{
si_bd_plt_rep(
  dat_rej,
  sel = NULL,
  type = c("zscore", "difference", "pvalue"),
  ref_line = TRUE
)
}
\description{
Plot data for all
}
