% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/si_boundary.R
\name{si_bd_plt_single}
\alias{si_bd_plt_single}
\title{Plot data for single trial}
\usage{
si_bd_plt_single(
  dat,
  type = c("y", "mean", "difference", "zscore", "pvalue"),
  ref_line = NULL,
  col = "black",
  n_first = NULL
)
}
\description{
Plot data for single trial
}
