% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/si_boundary.R
\name{si_bd_simu_trial}
\alias{si_bd_simu_trial}
\title{GSD}
\usage{
si_bd_simu_trial(
  n_tot = 200,
  rand_ratio = 1,
  trt_mean = 0,
  ctl_mean = 0,
  trt_sd = 1,
  ctl_sd = 1,
  ...,
  seed = 0
)
}
\description{
Functions for understanding group sequential design and interim analysis
boundaries
}
\keyword{Simulate}
\keyword{a}
\keyword{internal}
\keyword{single}
\keyword{trial}
