% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/si_door.R
\name{si_door_rank_bs}
\alias{si_door_rank_bs}
\title{DOOR Rank All with BS}
\usage{
si_door_rank_bs(
  dta_trt,
  dta_ctl,
  lst_outcomes,
  n_bs = 100,
  n_cores = 10,
  seed = NULL
)
}
\description{
DOOR Rank All with BS
}
