% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/si_simu.R
\name{si_simu_enroll}
\alias{si_simu_enroll}
\title{Simulate Enrollment Time}
\usage{
si_simu_enroll(
  ntot,
  enroll_duration = 12,
  min_fu = NULL,
  date_bos = NULL,
  mth_to_days = 30.4,
  round_days = FALSE,
  ...
)
}
\description{
Simulate Enrollment Time
}
