% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/si_zzz_package.R
\docType{package}
\name{statideas-package}
\alias{statideas-package}
\alias{statideas}
\title{The 'statideas' package.}
\description{
Demonstrate statistics ideas and concepts through interactive
    Shiny applications.
}
