% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stb_tls_gsd.R
\name{stb_tl_gsd_condpower}
\alias{stb_tl_gsd_condpower}
\title{Compute conditional power}
\usage{
stb_tl_gsd_condpower(
  zscore,
  info_frac,
  alpha = 0.025,
  power = 0.9,
  use_observed = TRUE
)
}
\description{
Formula is from Proschan, Lan and Wittes (2006) Chapter 3
}
