% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stb_tls_gsd.R
\name{stb_tl_gsd_predpower}
\alias{stb_tl_gsd_predpower}
\title{Compute predictive power}
\usage{
stb_tl_gsd_predpower(
  zscore,
  info_frac,
  alpha = 0.025,
  power = 0.9,
  prior_weight = 0.05
)
}
\description{
Formula is from Proschan, Lan and Wittes (2006) Chapter 3
}
