% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stb_tls_gsd.R
\name{stb_tl_gsd_prob}
\alias{stb_tl_gsd_prob}
\title{GSD boundaries}
\usage{
stb_tl_gsd_prob(
  info_fracs = c(0.5, 1),
  boundary = c(0.022, 0.022),
  boundary_type = c("alpha", "alpha_spent", "zscore"),
  alpha = 0.025,
  power = 0.9
)
}
\description{
GSD boundaries
}
