% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stb_tls_gsd.R
\name{stb_tl_gsd_rfun}
\alias{stb_tl_gsd_rfun}
\title{GSD boundaries by different R packages}
\usage{
stb_tl_gsd_rfun(
  typeOfDesign,
  info_fracs = c(0.5),
  alpha_spending = c(0.001),
  sided = 1,
  alpha = 0.025,
  rpackage = c("rpact", "gsDesign"),
  ...
)
}
\description{
GSD boundaries by different R packages
}
