% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stb_tls_gsd.R
\name{stb_tl_gsd_solve}
\alias{stb_tl_gsd_solve}
\title{GSD for one interim analysis}
\usage{
stb_tl_gsd_solve(
  info_fracs = c(0.2, 1),
  boundary = c(0.001, NA),
  boundary_type = c("alpha", "alpha_spent", "zscore"),
  alpha = 0.025,
  power = 0.9,
  ...
)
}
\description{
GSD for one interim analysis
}
